requirements.txt
autogen
python-dotenv
requests
openai
ag2[openai]
pyyaml
twilio
faiss-cpu
sentence-transformers